Moved encryption to plugins
