+++ server/plugins/encryption/type.go
